+++ src/main/java/com/stu/dto/EvaluationDTO.java
@@ -6,9 +6,9 @@ import java.util.List;
 
 @Data
 public class EvaluationDTO {
-    private Integer teacherId;
+    private Long teacherId;
 
-    private Float score;
+    private Double score;
 
 
     private String context;
